feat(unit-test): clean up package structure
