+++ main.go
@@ -1,6 +1,8 @@
-package utils
+package main
 
 import (
+	"fmt"
+	"io"
 	"os"
 	"strings"
 )
@@ -99,4 +101,72 @@ func CountWords(bytes []byte) int {
 	}
 
 	return wordCount
+}
+
+func main() {
+	var fileName string
+	var fileBytes []byte
+	var flag string
+
+	legalFlags := [4]string{"-c", "-l", "-w", "-m"}
+
+	file := os.Stdin
+	fileInfo, err := file.Stat()
+	if err != nil {
+		panic(err)
+	}
+
+	size := fileInfo.Size()
+	if size > 0 {
+		fileBytes, _ = io.ReadAll(os.Stdin)
+		if len(os.Args) > 1 {
+			for _, legalFlag := range legalFlags {
+				if os.Args[1] == legalFlag {
+					flag = os.Args[1]
+				}
+			}
+		}
+
+	} else {
+		if len(os.Args) < 3 {
+			for _, legalFlag := range legalFlags {
+				if os.Args[1] == legalFlag {
+					panic("Flag set with no filename")
+				}
+			}
+			fileName = os.Args[1]
+			fileBytes = FileToBytes(fileName)
+		} else {
+			for _, legalFlag := range legalFlags {
+				if os.Args[1] == legalFlag {
+					flag = os.Args[1]
+					fileName = os.Args[2]
+					fileBytes = FileToBytes(fileName)
+				}
+			}
+		}
+	}
+
+	if flag == "-c" {
+		fileByteTotal := CountBytes(fileBytes)
+		fmt.Println(fileByteTotal, fileName)
+
+	} else if flag == "-l" {
+		fileLines := CountLines(fileBytes)
+		fmt.Println(fileLines, fileName)
+
+	} else if flag == "-w" {
+		fileWords := CountWords(fileBytes)
+		fmt.Println(fileWords, fileName)
+
+	} else if flag == "-m" {
+		fileChars := CountChars(fileBytes)
+		fmt.Println(fileChars, fileName)
+
+	} else {
+		fileByteCount := CountBytes(fileBytes)
+		fileLines := CountLines(fileBytes)
+		fileWords := CountWords(fileBytes)
+		fmt.Println(fileLines, fileWords, fileByteCount, fileName)
+	}
 }
